Continue actuator v2.
The pipeline was refactored: there are now three parts:
1. Prepare the idf into an epjson with the correct meters.
2. Add controllable schedules to the epjson. Record the information of the
   created schedules.
3. Extract metadata from the epjson by doing a simulation.

Here is the complete lifetime (from zip to simulator) of a file from the
hydroquebec dataset:

1. Download zip (unchanged).
2. Extract zip (unchanged).
3. Index the contents (unchanged).
4. Prepare the idf. This steps upgrades the version of the file, it converts the
   format, sets up all the meters that will be necessary for creating the
   observation space during control and sets the correct timestep. This produces
   a file that is ready for control, except without the controllable schedules.
5. the schedule is linked into the epjson (just like before).
6. The epjson is made controllable. In the same step, a description of all the
   new actuators is also created. These actuator descriptions will be saved
   alongside the "ready for control" epjson and will be propagated up to the
   simulation step.
7. When a file is to be matched with a weather file and metadata to turn into a
   BuildingConfig, one additional step of the pipeline is used:
   extract_discovery_metadata. this function modifies the epjson such that it
   produces the necessary output files and runs a simulation in order to extract
   relevant metadata such as the controlled area and the number of warmup
   phases. That metadata is combined with the epjson, the epw and the actuator
   descriptions to form a BuildingConfig.
8. From that point on, the BuildingConfig lives its life as usual.

+++ building2building/pipeline/parse_reports.py
@@ -56,7 +56,7 @@ def get_net_conditioned_area(html_path: Path) -> float:
     return value
 
 
-def get_warmup_days(html_path: Path) -> float:
+def get_warmup_days(html_path: Path) -> int:
     """Extract the number of warm-up days from an EnergyPlus HTML summary file."""
     with open(html_path, "r", encoding="utf-8", errors="ignore") as f:
         soup = BeautifulSoup(f, "html.parser")
@@ -111,6 +111,5 @@ def get_warmup_days(html_path: Path) -> float:
 
     if series.empty:
         raise Exception("could not read warmup days")
-    warmup_days = float(series.iloc[0])
+    warmup_days = int(series.iloc[0])
     return warmup_days
-
